+++ build-config.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-	extraFiles: ['assets/fonts/Roboto-Regular.ttf', 'assets/fonts/Roboto-Bold.ttf'],
-}